+++ src/Order/OrderService.php
@@ -14,6 +14,7 @@ use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
+use Wizaplace\SDK\Subscription\SubscriptionSummary;
 
 /**
  * Class OrderService
@@ -283,5 +284,15 @@ final class OrderService extends AbstractService
             }
             throw $e;
         }
+    }
+
+    /** @return SubscriptionSummary[] */
+    public function getSubscriptions(int $orderId): array
+    {
+        $this->client->mustBeAuthenticated();
+
+        return array_map(function (array $data): SubscriptionSummary {
+            return new SubscriptionSummary($data);
+        }, $this->client->get("user/orders/${orderId}/subscriptions"));
     }
 }